@@ -2,12 +2,13 @@
 
 from .fix_message import FixMessage, FixMessageFactory
 from .loader import load_yaml_protocol, load_quickfix_protocol
-from .types import ValueType
+from .types import ValueType, StrictMode
 
 __all__ = [
     'FixMessage',
     'FixMessageFactory',
     'load_quickfix_protocol',
     'load_yaml_protocol',
-    'ValueType'
+    'ValueType',
+    'StrictMode'
 ]
