@@ -1,5 +1,6 @@
-"""A loader for YAML formated metadata"""
+"""A loader for YAML formatted metadata"""
 
+from pathlib import Path
 from typing import Mapping, Optional, Union
 
 from ruamel.yaml import YAML
@@ -11,7 +12,7 @@ from .loader import load_protocol
 
 
 def load_yaml_protocol(
-        filename: str,
+        filename: Union[str, Path],
         *,
         is_millisecond_time: bool = True,
         is_float_decimal: bool = False,
@@ -20,7 +21,7 @@ def load_yaml_protocol(
     """Load a YAML style protocol file
 
     Args:
-        filename (str): The filename
+        filename (Union[str, Path]): The filename
         is_millisecond_time (bool, optional): If true times have milliseconds.
             Defaults to True.
         is_float_decimal (bool, optional): If true use Decimal for floating
@@ -31,8 +32,12 @@ def load_yaml_protocol(
     Returns:
         ProtocolMetaData: The protocol meta data.
     """
+    if not isinstance(filename, Path):
+        filename = Path(filename)
+
     yaml = YAML()
-    with open(filename, 'rt') as file_ptr:
+
+    with filename.open('rt', encoding="utf8") as file_ptr:
         return load_protocol(
             yaml.load(file_ptr),
             is_millisecond_time=is_millisecond_time,
